pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gai-cli"
version = "0.1.0"
description = "A production-ready Python CLI tool for Google AI Studio."
readme = "README.md"
authors = [
  { name = "Bugra", email = "bugra@example.com" },
]
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.10"
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "google-genai>=0.1.0",
    "prompt_toolkit>=3.0.0",
]

[project.urls]
"Homepage" = "https://github.com/bugraakdemir/gai-cli"
"Bug Tracker" = "https://github.com/bugraakdemir/gai-cli/issues"

[project.scripts]
gai = "gai.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
